[REFACTOR] #18 - BaseActivity 수정
